class refactoring update
