+++ Dockerfile
@@ -7,4 +7,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 COPY . .
 
-CMD ["python", "bot/bot.py"]
+CMD ["python", "bot/main.py"]
